==> 数据生成/2.barra_factor_generator.ipynb <==
"""
生成Barra因子所需的各个描述因子
数据来自wind和聚宽两个渠道，其中一致预期类数据有大量缺失，效果不好
"""

import pandas as pd
import numpy as np
import akshare as ak
import statsmodels.formula.api as smf
import statsmodels.api as sm
from sklearn.preprocessing import StandardScaler
from scipy.stats.mstats import winsorize
import datetime
import math
import matplotlib.pyplot as plt
from sklearn.linear_model import LinearRegression

"""
数据导入和整理
"""

dataset_keys = [
    '涨跌幅(%)',
    'R_f',
    'R_m',
    'market_cap',  # 总市值
    '换手率(%)',
    '市净率',
    '市盈率',
    'total_liability',
    'total_assets',
    'equities_parent_company_owners', # 归属于母公司股东权益合计
    'preferred_shares_equity',
    'total_non_current_liability',
    'A股流通市值(元)',
    'FY1',
    'FY3',
    'FY12',
    'pcf_ratio',
    'operating_revenue',
    '总股本(股)',
    'eps',
    'circulating_market_cap',
    'net_profit',
    '总市值(元)',
    '收盘价(元)',
]

# 读取数据
dataset_dict = {}
for key in dataset_keys : 
    dataset_dict[key] = pd.read_csv('dataset/data_of_factor_need/' + key + '.csv', low_memory=False).set_index('date')
    
# 读取tradingdate，并获取对应一年前的交易日日期
trading_date_list = pd.read_csv('dataset/trading_date.csv').iloc[252:,:] 
dates = trading_date_list.loc[252*2:3000,'date']
def last_year_func(x):
    index_x = dates[dates.values==x].index
    try:
        return dates[index_x - 252].values[0]
    except : 
        return np.nan
dates_last = dates.apply(last_year_func)  

# 读取股票列表stks，这里用来自wind和来自jqdata的数据的交集
stks = list(set(dataset_dict['涨跌幅(%)'].columns) & (set(dataset_dict['FY3'].columns)) & (set(dataset_dict['pcf_ratio'].columns) ))
stks.sort()
dates = sorted(list(set(dates) & set(dataset_dict['pcf_ratio'].index)))

## Beta

def beta(stks, dates):
    # '000001.SZ'
    T = 252
    L = 0
    half_day = 63
    weights = 0.5**(np.arange(252-1,-1,-1)/63) 
    weights = weights / sum(weights)
    dates_i = dates[252:]
    def beta_inner(date):
        index_x = dates[dates.values==date].index
        date_last = dates[index_x - 251].values[0]
        
        stks_data = dataset_dict['涨跌幅(%)'].loc[date_last:date,stks].sort_index(axis=0, ascending=True)
        rf_data = dataset_dict['R_f'].loc[date_last:date,'r_f_daily'].sort_index(axis=0, ascending=True)
        rm_data = dataset_dict['R_m'].loc[date_last:date,'market'].sort_index(axis=0, ascending=True)
        
        y = stks_data.sub(rf_data, axis="index")
        x = rm_data.sub(rf_data, axis="index")

        sum_w_x_y = ((y.sub(y.mean(axis=1),axis=0)).multiply((x - x.mean()),axis=0).multiply(weights,axis=0).mean() * len(y))
        sum_w_x2 = ((x - x.mean())**2).multiply(weights,axis=0).sum()
        
#         alpha_value = np.mean(y) - beta_value * np.mean(x) 

#         ei = y.sub(alpha_value.add(beta_value.mul(x,axis=0)))
        return sum_w_x_y / sum_w_x2
    dates_i.index = dates_i
    aa = dates_i.apply(beta_inner)    
    
    return aa

beta(stks, dates).to_csv('dataset/barra_factor/' + 'beta' + '.csv')

## Size

def lncap(stks, dates):
    # '000001.SZ'
    return np.log(dataset_dict['总市值(元)'].loc[dates,stks])

lncap(stks, dates).to_csv('dataset/barra_factor/' + 'lncap' + '.csv')

## Residual Volatility

$$dastd = \frac{1}{n} \sum_{t=1}^{n}{w_t {(r_{et} - \overline{r_e})}^2}$$
$$ r_{et} = r_t - r_{ft} $$

def dastd(dates):
    # '000001.SZ'
    T = 252
    L = 0
    half_day = 42
    
    dates_i = dates[252:]
    def beta_inner(date):
        index_x = dates[dates.values==date].index
        date_last = dates[index_x - 252].values[0]
        
        stks_data = dataset_dict['涨跌幅(%)'].loc[date_last:date,stks].sort_index(axis=0, ascending=True)
        rf_data = dataset_dict['R_f'].loc[date_last:date,'r_f_daily'].sort_index(axis=0, ascending=True)
        rm_data = dataset_dict['R_m'].loc[date_last:date,'market'].sort_index(axis=0, ascending=True)
        
        ret = stks_data.sub(rf_data, axis="index")
        ret_mean = np.mean(ret)

        return (pd.DataFrame.ewm((ret - ret_mean)**2, halflife=half_day,adjust=False).mean().iloc[-1]) / (T - L)

    dates_i.index = dates_i
    aa = dates_i.apply(beta_inner)    
    
    return aa
(dastd(dates)*10000).to_csv('dataset/barra_factor/' + 'dastd' + '.csv')

$$ Z(T) = \sum_{t=1}^{T}{[ln(1+r_t)-ln(1+r_{ft})]} $$
$$crma = ln(1+Z_{max})-ln(1+Z_{min})$$

def cmra(stks, dates) : 
    stks_chg_data = pd.concat([dataset_dict['R_f']['r_f_daily'], dataset_dict['涨跌幅(%)'][stks]],axis=1).sort_index()
    T_month = [252 - 21 * month for month in range(13)]
    dates_i = dates[252:]
    def inner(date):
        index_loc = stks_chg_data.index.get_loc(date)
#         print(date, index_loc)
        T_month_date = index_loc - np.array(T_month)
#         print(T_month_date)
        def inner_a(i) : 
            # 计算Z_T(未累加)
            i = i.values[0]
            start_day = T_month_date[i]
            end_day = T_month_date[i + 1]
            ln_rf_rft = np.log(1+stks_chg_data.iloc[start_day:end_day, 1:]).\
                    sub(np.log(1+stks_chg_data.iloc[start_day:end_day, 0]),axis=0)  
            ln_rf_rft.loc[:,ln_rf_rft.isnull().sum()/len(ln_rf_rft) > 0.3] = np.nan     #   股票21天内停牌时间不超过30%
            res = ln_rf_rft.mean() * len(stks_chg_data.iloc[start_day:end_day, :])   
            res = res.rename(end_day)
            return res
        innera = pd.DataFrame([i for i in range(12)],index=[i for i in range(12)])
        innera = innera.apply(inner_a,axis=1) 
        innera.loc[:,innera.isnull().sum()/len(innera) > 0.5] = np.nan  # 股票上市需要超过六个月
        innera = innera.iloc[::-1,:].cumsum().iloc[::-1,:] # 累加以获得Z_T
        return np.log(1 + np.max(innera)) - np.log(1 + np.min(innera))  

    dates_i.index = dates_i
    aa = dates_i.apply(inner)    
    
    return aa
res = cmra(stks, dates)
res.to_csv('dataset/barra_factor/' + 'cmra' + '.csv')

$$ hsigma = std(e_t) $$

def hsigma(stks, dates):
    T = 252
    L = 0
    half_day = 63
    
    dates_i = dates[252:]
    def beta_inner(date):
        index_x = dates[dates.values==date].index
        date_last = dates[index_x - 252].values[0]
        
        stks_data = dataset_dict['涨跌幅(%)'].loc[date_last:date,stks].sort_index(axis=0, ascending=True)
        rf_data = dataset_dict['R_f'].loc[date_last:date,'r_f_daily'].sort_index(axis=0, ascending=True)
        rm_data = dataset_dict['R_m'].loc[date_last:date,'market'].sort_index(axis=0, ascending=True)
        
        y = stks_data.sub(rf_data, axis="index")
        x = rm_data.sub(rf_data, axis="index")

        cov = (y - np.mean(y)).multiply(x - np.mean(x), axis="index")
        doc = (x - np.mean(x))**2

        a = pd.DataFrame.ewm(cov,halflife=half_day,adjust=False).mean().iloc[-1]
        b = pd.DataFrame.ewm(doc,halflife=half_day,adjust=False).mean().iloc[-1] 
        beta_value = a/b

        alpha_value = np.mean(y) - (beta_value) * np.mean(x) 

        ei = y.sub(alpha_value.add(pd.DataFrame([beta_value]*len(x),index=x.index).mul(x,axis=0)))
        return ei.std()
    dates_i.index = dates_i
    aa = dates_i.apply(beta_inner)    
    
    return aa

hsigma(stks, dates).to_csv('dataset/barra_factor/' + 'hsigma' + '.csv')

## Momentum

$$rstr =  \sum_{t=L}^{T+L}{w_t ln(1+r_t)} - \sum_{t=L}^{T+L}{w_t ln(1+r_{ft})}$$

def rstr(dates):
    half_day = 126
    dates_i = dates[504+21:]
    def rstr_inner(date):
        weights = 0.5**(np.arange(504-1,-1,-1)/126)   #指数衰减权重 # new: T to h
        index_x = dates[dates.values==date].index
        date = dates[index_x - 21].values[0]
        date_last = dates[index_x - 504 - 20].values[0]
        rt = dataset_dict['涨跌幅(%)'].loc[date_last:date,stks].sort_index(axis=0, ascending=True)
        rft = dataset_dict['R_f'].loc[date_last:date,'r_f_daily'].sort_index(axis=0, ascending=True)
        
        return (rt.multiply(weights, axis=0)).mean() - (rft.multiply(weights, axis=0)).mean()
    dates_i.index = dates_i
    aa = dates_i.apply(rstr_inner)    
    return aa
res_df = rstr(dates)
res_df.to_csv('dataset/barra_factor/' + 'rstr' + '.csv')


## Liquidity

$$stom=ln(\sum^{21}_{t=1}{\frac{V_t}{S_t}})$$

stks_stom_data = pd.read_csv('dataset/后复权数据-分类/换手率(%).csv',encoding='gbk')
stks_stom_data = stks_stom_data.rename(columns = {"日期": "date"}).set_index('date')
stks_stom_data_columns = stks_stom_data.columns.to_list()
for i in range(len(stks_stom_data_columns)) : 
    if stks_stom_data_columns[i][:1] == '6' : 
        stks_stom_data_columns[i] = stks_stom_data_columns[i][:6] + '.XSHG'
    else : 
        stks_stom_data_columns[i] = stks_stom_data_columns[i][:6] + '.XSHE'
stks_stom_data.columns = stks_stom_data_columns
stks_stom_data = stks_stom_data[stks].sort_index()

def stom(date):
    T = 21
    L = 0
    dates_i = dates[21:]
    def stom_inner(date):
        index_x = dates[dates.values==date].index
        date_last = dates[index_x - 21].values[0]
        stks_stom_data = dataset_dict['换手率(%)'].loc[date_last:date,stks]
        
        return np.log(stks_stom_data.sum())
    dates_i.index = dates_i
    aa = dates_i.apply(stom_inner)    
    return aa

stom_data = stom(dates)
stom_data = stom_data.replace([np.inf, -np.inf], np.nan)
stom_data.to_csv('dataset/barra_factor/' + 'stom' + '.csv')

$$stoq=ln(\frac{1}{3}\sum^{3}_{t=1}{exp(stom_t)})$$

    # '000001.SZ'
    T = 21 * 3
    L = 0
    dates = pd.Series(stom_data.index.to_list())
    dates_i = dates[T: ]
    def stoq_inner(date):
        index_x = dates[dates.values==date].index
        date_last = dates[index_x - T].values[0]
        stks_stom_data = stom_data.loc[date_last:date,stks]
        
        return np.log(stks_stom_data.sum()/3)
    dates_i.index = dates_i
    aa = dates_i.apply(stoq_inner)    
    return aa
stoq_df = stoq()
stoq_df = stoq_df.replace([np.inf, -np.inf], np.nan)
stoq_df.to_csv('dataset/barra_factor/' + 'stoq' + '.csv')

$$stoa=ln(\frac{1}{12}\sum^{12}_{1}{exp(stom_t)})$$

def stoa():
    T = 21 * 12
    L = 0
    dates = pd.Series(stom_data.index.to_list())
    dates_i = dates[T: ]
    def stoa_inner(date):
        index_x = dates[dates.values==date].index
        date_last = dates[index_x - T].values[0]
        stks_stom_data = stom_data.loc[date_last:date,stks]
        
        return np.log(stks_stom_data.sum()/12)
    dates_i.index = dates_i
    aa = dates_i.apply(stoa_inner)    
    return aa
stoa_df = stoa()
stoa_df = stoa_df.replace([np.inf, -np.inf], np.nan)
stoa_df.to_csv('dataset/barra_factor/' + 'stoa' + '.csv')

## Non Linear Size

def nlsize(stks, date):
    size_factor = np.log(dataset_dict['总市值(元)'].loc[dates,stks])
    size_factor_3 = size_factor**3
    dates_i = dates[252: ]
    
    def ols_inner(date):
        aa = (dataset_dict['总市值(元)'].loc[date,stks].dropna()) / sum(dataset_dict['总市值(元)'].loc[date,stks].dropna())
        WLS = LinearRegression()
        WLS.fit(pd.DataFrame(size_factor_3.loc[date].dropna()), pd.DataFrame(dataset_dict['总市值(元)'].loc[date,stks].dropna()), 
                    sample_weight = aa)
        e_value = (pd.DataFrame(dataset_dict['总市值(元)'].loc[date,stks].dropna()) - WLS.predict(pd.DataFrame(size_factor_3.loc[date].dropna())) ).T
        e_value[list(set(stks) - set(e_value.columns))] = np.nan
        e_value = (e_value[stks])
        return e_value.iloc[0]
    dates_i.index = dates_i
    aa = dates_i.apply(ols_inner)    
    
    return aa
    
nlsize(stks, dates).to_csv('dataset/barra_factor/' + 'nlsize' + '.csv')

## Book to price

def btop(stks, dates) : 
    bp_ratio = dataset_dict['收盘价(元)'].loc[dates[252:],stks] * \
                dataset_dict['总股本(股)'].loc[dates[252:],stks] / \
                dataset_dict['equities_parent_company_owners'].loc[dates[252:],stks]
    res = 1/(bp_ratio)
    res[res<0] = np.nan
    return res
btop(stks, dates).to_csv('dataset/barra_factor/' + 'btop' + '.csv')

## Earning yeild

def epfwd(): 
    return dataset_dict['FY12'].loc[dates[252:],stks] / dataset_dict['market_cap'].loc[dates[252:],stks]

epfwd().to_csv('dataset/barra_factor/' + 'epfwd' + '.csv')

def cetop(): 
    return 1/(dataset_dict['pcf_ratio'].loc[dates[252:],stks] / dataset_dict['总股本(股)'].loc[dates[252:],stks])

cetop().to_csv('dataset/barra_factor/' + 'cetop' + '.csv')

def etop(): 
    return 1 / dataset_dict['市盈率'].loc[dates[252:],stks]

etop().to_csv('dataset/barra_factor/' + 'etop' + '.csv')

## Growth

def egrlf(): 
    return dataset_dict['FY3'].loc[dates[252:],stks] / dataset_dict['net_profit'].loc[dates[252:],stks]
egrlf().to_csv('dataset/barra_factor/' + 'egrlf' + '.csv')

def egrsf(): 
    return dataset_dict['FY1'].loc[dates[252:],stks] / dataset_dict['net_profit'].loc[dates[252:],stks]
egrsf().to_csv('dataset/barra_factor/' + 'egrsf' + '.csv')

def egro(dates): 
    data_df = dataset_dict['eps'].loc[dates[252:],stks]
    dates_i = dates[252: ]
    def ols_inner(date):
        index_loc = data_df.index.get_loc(date)
        ols_list = []
        for i in range(5) : 
            ols_list.append(index_loc - 252*i)
        y = data_df.iloc[ols_list,:]
        x = [1,2,3,4,5]

        cov = (y - np.mean(y)).multiply(x - np.mean(x), axis="index")
        doc = (x - np.mean(x))**2
        
        a = cov.sum()
        b = doc.sum()
        beta_value = a/b
        return beta_value
    dates_i.index = dates_i
    aa = dates_i.apply(ols_inner)    
    
    return aa
egro(dates).to_csv('dataset/barra_factor/' + 'egro' + '.csv')

def sgro(dates): 
    data_df = dataset_dict['operating_revenue'].loc[dates[252:],stks] / dataset_dict['总股本(股)'].loc[dates[252:],stks]
    dates_i = dates[252: ]
    def ols_inner(date):
        index_loc = data_df.index.get_loc(date)
        ols_list = []
        for i in range(5) : 
            ols_list.append(index_loc - 252*i)
        y = data_df.iloc[ols_list,:]
        x = [1,2,3,4,5]

        cov = (y - np.mean(y)).multiply(x - np.mean(x), axis="index")
        doc = (x - np.mean(x))**2
        
        a = cov.sum()
        b = doc.sum()
        beta_value = a/b
        return beta_value
    dates_i.index = dates_i
    aa = dates_i.apply(ols_inner)    
    return aa
sgro(dates).to_csv('dataset/barra_factor/' + 'sgro' + '.csv')

## Leverage

$$ mlev = \frac{me+pe+ld}{me}$$
mlev: 市场杠杆,me: 普通股市值,pe: 优先股账面价值,ld: 长期负债账面价值

def mlev(stks, dates):
    result = (dataset_dict['A股流通市值(元)'].loc[dates[252:],stks] + \
            dataset_dict['preferred_shares_equity'].loc[dates[252:],stks].fillna(0) + \
            dataset_dict['total_non_current_liability'].loc[dates[252:],stks].fillna(0)) / \
            dataset_dict['A股流通市值(元)'].loc[dates[252:],stks]
    return result


mlev(stks, dates).to_csv('dataset/barra_factor/' + 'mlev' + '.csv')

$$ dtoa = \frac{td}{ta}$$
dtoa: 资产负债比,td: 总负债账面价值,ta: 总资产账面价值

def dtoa(stks, dates) : 
    # new
    return dataset_dict['total_liability'].loc[dates[252:],stks] / dataset_dict['total_assets'].loc[dates[252:],stks]
dtoa(stks, dates).to_csv('dataset/barra_factor/' + 'dtoa' + '.csv')

$$ blev = \frac{be+pe+ld}{be}$$
blev: 账面杠杆,be: 普通股账面价值,pe: 优先股账面价值,ld: 长期负债账面价值

def blev(stks, date) : 
    result = (dataset_dict['equities_parent_company_owners'].loc[dates[252:],stks] + \
            dataset_dict['preferred_shares_equity'].loc[dates[252:],stks].fillna(0) + \
            dataset_dict['total_non_current_liability'].loc[dates[252:],stks].fillna(0)) / \
            dataset_dict['equities_parent_company_owners'].loc[dates[252:],stks]
    return result
blev_df = blev(stks, dates)
blev_df[(blev_df<0)&(blev_df>100)] = np.nan
blev_df.to_csv('dataset/barra_factor/' + 'blev' + '.csv')
